requirements.txt
pygame-ce >= 2.5.3
numpy >= 2.2.6
numba >= 0.61.2
scipy >= 1.15.3
fastapi >= 0.115.12
python-multipart >= 0.0.20
uvicorn >= 0.34.2
opencv-python >= 4.11.0.86
tqdm >= 4.67.1
pillow >= 11.2.1
lapx



torch~=2.7.1
torchvision~=0.22.1
matplotlib~=3.10.3